refactor(compat-corpus): extract shared svelte-check diagnostic plumbing
The Layer 2 e2e gate compares the same thing as `check-verify.mjs` — the
diagnostic set official svelte-check and rsvelte-check report for one
type-checked project — so both must normalize and diff it identically.
Move the machine-verbose parser, the diagnostic key, the multiset diff and
the exit-tolerant runner into `check-diagnostics.mjs`; `check-verify.mjs`
behaviour is unchanged (Layer 1 still reports 0 divergences).
